# apps/backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/branches/{branchId}/tables": {
      "post": {
        "operationId": "TableController_createTable",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTableDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TableResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create table",
        "tags": [
          "Tables & Reservations"
        ]
      },
      "get": {
        "operationId": "TableController_findAllTables",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TableResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List tables for a branch",
        "tags": [
          "Tables & Reservations"
        ]
      }
    },
    "/tables/{id}": {
      "patch": {
        "operationId": "TableController_updateTable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTableDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TableResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update table",
        "tags": [
          "Tables & Reservations"
        ]
      },
      "delete": {
        "operationId": "TableController_deleteTable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete table",
        "tags": [
          "Tables & Reservations"
        ]
      }
    },
    "/branches/{branchId}/reservations": {
      "post": {
        "operationId": "TableController_createReservation",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReservationResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create reservation",
        "tags": [
          "Tables & Reservations"
        ]
      },
      "get": {
        "operationId": "TableController_findAllReservations",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "CONFIRMED",
                "SEATED",
                "COMPLETED",
                "CANCELLED",
                "NO_SHOW"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ReservationResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List reservations for a branch",
        "tags": [
          "Tables & Reservations"
        ]
      }
    },
    "/reservations/{id}": {
      "patch": {
        "operationId": "TableController_updateReservation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReservationResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update reservation",
        "tags": [
          "Tables & Reservations"
        ]
      }
    },
...
  },
  "info": {
    "title": "milky",
    "description": "The milky API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "MetaDataDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "total": {
            "type": "number",
            "example": 50
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
      "GenericResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDataDto"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "TableResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "serviceStatus": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "OCCUPIED",
              "ORDERING",
              "PREPARING",
              "SERVING",
              "PAYMENT_PENDING",
              "CLOSING"
            ]
          },
          "operationalFlag": {
            "type": "string",
            "enum": [
              "NORMAL",
              "RESERVED",
              "CLEANING",
              "OUT_OF_SERVICE"
            ]
          },
          "qrCodeToken": {
            "type": "string"
          },
          "shape": {
            "type": "string",
            "enum": [
              "CIRCLE",
              "SQUARE",
              "RECTANGLE",
              "TRIANGLE",
              "HEXAGON"
            ]
          },
          "floor": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "branchId",
          "label",
          "capacity",
          "serviceStatus",
          "operationalFlag",
          "shape",
          "floor"
        ]
      },
      "CreateTableDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "minimum": 1
          },
          "shape": {
            "type": "string",
            "enum": [
              "CIRCLE",
              "SQUARE",
              "RECTANGLE",
              "TRIANGLE",
              "HEXAGON"
            ],
            "default": "SQUARE"
          },
          "floor": {
            "type": "number",
            "default": 1
          }
        },
        "required": [
          "label",
          "capacity"
        ]
      },
      "UpdateTableDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "minimum": 1
          },
          "shape": {
            "type": "string",
            "enum": [
              "CIRCLE",
              "SQUARE",
              "RECTANGLE",
              "TRIANGLE",
              "HEXAGON"
            ],
            "default": "SQUARE"
          },
          "floor": {
            "type": "number",
            "default": 1
          },
          "serviceStatus": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "OCCUPIED",
              "ORDERING",
              "PREPARING",
              "SERVING",
              "PAYMENT_PENDING",
              "CLOSING"
            ]
          },
          "operationalFlag": {
            "type": "string",
            "enum": [
              "NORMAL",
              "RESERVED",
              "CLEANING",
              "OUT_OF_SERVICE"
            ]
          }
        }
      },
      "ReservationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          },
          "tableId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "tableSessionId": {
            "type": "string"
          },
          "guestName": {
            "type": "string"
          },
          "guestPhone": {
            "type": "string"
          },
          "partySize": {
            "type": "number"
          },
          "reservedFor": {
            "format": "date-time",
            "type": "string"
          },
          "durationMins": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "SEATED",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW"
            ]
          },
          "specialRequest": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "branchId",
          "guestName",
          "guestPhone",
          "partySize",
          "reservedFor",
          "durationMins",
          "status"
        ]
      },
      "CreateReservationDto": {
        "type": "object",
        "properties": {
          "tableId": {
            "type": "string"
          },
          "guestName": {
            "type": "string"
          },
          "guestPhone": {
            "type": "string"
          },
          "partySize": {
            "type": "number",
            "minimum": 1
          },
          "reservedFor": {
            "type": "string"
          },
          "durationMins": {
            "type": "number",
            "default": 90
          },
          "specialRequest": {
            "type": "string"
          }
        },
        "required": [
          "guestName",
          "guestPhone",
          "partySize",
          "reservedFor"
        ]
      },
      "UpdateReservationDto": {
        "type": "object",
        "properties": {
          "tableId": {
            "type": "string"
          },
          "guestName": {
            "type": "string"
          },
          "guestPhone": {
            "type": "string"
          },
          "partySize": {
            "type": "number",
            "minimum": 1
          },
          "reservedFor": {
            "type": "string"
          },
          "durationMins": {
            "type": "number",
            "default": 90
          },
          "specialRequest": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "SEATED",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW"
            ]
          }
        }
      },
...
    }
  }
}